refactor: add component OrderSummaryItem and reuse it in OrderSummary component
