--- src/features/Order/index.ts
@@ -1,1 +1,0 @@
-export { OrderSummary } from "./ui/Order";
