+++ src/features/OrderSummary/index.ts
@@ -1,0 +1,1 @@
+export { OrderSummary } from "./ui/OrderSummary";
